Added in the PMT holder extra width and depth parameters and the corresponding print statements to match equations.txt
Other updates to equations.txt include swapping order of trans-open-closed to open-trans-closed wherever it appears - this means 4-5 swapped to 5-4, and I made other aesthetic changes.

There could be errors between this and the GDML generator now, but I tried to avoid messing with the GDML generator at all - I did fix the relative sign error - so all angles should be defined as positive in cadp.csv to get CAD to work, and GDML too.