Added a new field to detectorMotherP.csv for front/back extension
Without the last field, there is not extension to the front or
the back of the main mother volume. With a positive value, the
mother is extended at the back; with a negative value at the front.

This allows a mother volume that more closely ressembles the actual
shape of the detector array so as to fit inside the narrow space of
the actual experiment.

Note: The spaceship operator ($x<=>0) returns the sign of x in perl.